+++ SKPhotoBrowser/SKButtons.swift
@@ -22,8 +22,8 @@ class SKButton: UIButton {
         }
     }
     var size: CGSize = CGSize(width: 44, height: 44)
-    var margin: CGFloat = 5
-    var buttonTopOffset: CGFloat { return 5 }
+    var margin: CGFloat = SKPhotoBrowserOptions.closeAndDeleteButtonPadding
+    var buttonTopOffset: CGFloat { return 30 }
 
     func setup(_ imageName: String) {
         backgroundColor = .clear
@@ -42,10 +42,24 @@ class SKButton: UIButton {
         showFrame = newRect
         hideFrame = CGRect(x: margin, y: -20, width: size.width, height: size.height)
     }
+
+    func updateFrame(_ frameSize: CGSize) { }
 }
 
-class SKCloseButton: SKButton {
-    let imageName = "btn_common_close_wh"
+class SKImageButton: SKButton {
+
+    fileprivate var leftSidePositionMargin: CGFloat {
+        return super.margin
+    }
+
+    fileprivate var rightSidePositionMargin: CGFloat {
+        return SKMesurement.screenWidth - super.margin - self.size.width
+    }
+
+    fileprivate var imageName: String {
+        return ""
+    }
+
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
     }
@@ -58,10 +72,18 @@ class SKCloseButton: SKButton {
     }
 }
 
-class SKDeleteButton: SKButton {
-    let imageName = "btn_common_delete_wh"
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
+class SKCloseButton: SKImageButton {
+    override var imageName: String { return "btn_common_close_wh" }
+
+    override var margin: CGFloat {
+        get {
+            return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
+                ? rightSidePositionMargin
+                : leftSidePositionMargin
+        }
+        set {
+            super.margin = newValue
+        }
     }
 
     override init(frame: CGRect) {
@@ -73,6 +95,10 @@ class SKDeleteButton: SKButton {
         hideFrame = CGRect(x: SKMesurement.screenWidth - size.width, y: -20, width: size.width, height: size.height)
     }
 
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
     override func setFrameSize(_ size: CGSize) {
         let newRect = CGRect(x: SKMesurement.screenWidth - size.width, y: buttonTopOffset,
                              width: size.width,
@@ -80,5 +106,28 @@ class SKDeleteButton: SKButton {
         self.frame = newRect
         showFrame = newRect
         hideFrame = CGRect(x: SKMesurement.screenWidth - size.width, y: -20, width: size.width, height: size.height)
+    }
+}
+
+class SKDeleteButton: SKImageButton {
+    override var imageName: String { return "btn_common_delete_wh" }
+
+    override var margin: CGFloat {
+        get { return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
+            ? leftSidePositionMargin
+            : rightSidePositionMargin
+        }
+        set { super.margin = newValue }
+    }
+
+    override func updateFrame(_ newScreenSize: CGSize) {
+        showFrame = CGRect(x: newScreenSize.width - size.width,
+                           y: buttonTopOffset,
+                           width: size.width,
+                           height: size.height)
+        hideFrame = CGRect(x: newScreenSize.width - size.width,
+                           y: -20,
+                           width: size.width,
+                           height: size.height)
     }
 }